# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controller.SuccessResponse:
    properties:
      result:
        type: string
    type: object
  domain.AccountRateResponse:
    properties:
      account_id:
        type: string
      icon:
        type: string
      rate:
        type: integer
      username:
        type: string
    type: object
...
  domain.AddChatRequestBody:
    properties:
      account_id:
        type: string
      message:
        type: string
    required:
    - account_id
    - message
    type: object
...
  domain.CreateRateRequestBody:
    properties:
      rate:
        type: integer
    type: object
  domain.CreateRoomRequestBody:
    properties:
      account_id:
        type: string
      description:
        type: string
      hackathon_id:
        type: integer
      member_limit:
        type: integer
      title:
        type: string
    required:
    - account_id
    - description
    - hackathon_id
    - member_limit
    - title
    type: object
...
  domain.GetRoomResponse:
    properties:
      create_at:
        type: string
      description:
        type: string
      hackathon:
        $ref: '#/definitions/domain.RoomHackathonInfo'
      is_status:
        type: boolean
      member_limit:
        type: integer
      members_frameworks:
        items:
          $ref: '#/definitions/domain.RoomFramework'
        type: array
      members_tech_tags:
        items:
          $ref: '#/definitions/domain.RoomTechTags'
        type: array
      now_member:
        items:
          $ref: '#/definitions/domain.NowRoomAccounts'
        type: array
      room_id:
        type: string
      title:
        type: string
    type: object
...
  domain.ListRoomHackathonInfo:
    properties:
      hackathon_id:
        type: integer
      hackathon_name:
        type: string
      icon:
        type: string
    type: object
  domain.ListRoomResponse:
    properties:
      hackathon:
        $ref: '#/definitions/domain.ListRoomHackathonInfo'
      members_frameworks:
        items:
          $ref: '#/definitions/domain.RoomFramework'
        type: array
      members_tech_tags:
        items:
          $ref: '#/definitions/domain.RoomTechTags'
        type: array
      now_member:
        items:
          $ref: '#/definitions/domain.NowRoomAccounts'
        type: array
      rooms:
        $ref: '#/definitions/domain.ListRoomRoomInfo'
    type: object
  domain.ListRoomRoomInfo:
    properties:
      created_at:
        type: string
      member_limit:
        type: integer
      room_id:
        type: string
      title:
        type: string
    type: object
  domain.NowRoomAccounts:
    properties:
      account_id:
        type: string
      icon:
        type: string
      is_owner:
        type: boolean
    type: object
...
  domain.RoomFramework:
    properties:
      count:
        type: integer
      framework:
        $ref: '#/definitions/repository.Framework'
    type: object
  domain.RoomHackathonInfo:
    properties:
      hackathon_id:
        type: integer
      icon:
        type: string
      link:
        type: string
      name:
        type: string
      start_date:
        type: string
      status_tag:
        items:
          $ref: '#/definitions/repository.StatusTag'
        type: array
      term:
        type: integer
    type: object
  domain.RoomTechTags:
    properties:
      count:
        type: integer
      tech_tag:
        $ref: '#/definitions/repository.TechTag'
    type: object
...
  domain.UpdateRoomRequestBody:
    properties:
      description:
        type: string
      hackathonID:
        type: integer
      member_limit:
        type: integer
      title:
        type: string
    type: object
...
  repository.Framework:
    properties:
      framework:
        type: string
      framework_id:
        type: integer
      tech_tag_id:
        type: integer
    type: object
...
  repository.StatusTag:
    properties:
      status:
        type: string
      status_id:
        type: integer
    type: object
  repository.TechTag:
    properties:
      icon:
        type: string
      language:
        type: string
      tech_tag_id:
        type: integer
    type: object
host: https://seaffood.com
info:
  contact:
    email: oogiriminister@gmail.com
    name: murasame
    url: https://twitter.com/fresh_salmon256
  description: HackPortal Backend API serice
  license:
    name: No-license
    url: No-license
  termsOfService: https://seaffood.com/api
  title: Hack Hack Backend API
  version: "1.0"
paths:
...
  /accounts/{account_id}/rate:
    get:
      description: List Rate for User
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      - in: query
        name: page_id
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/domain.AccountRateResponse'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List Rate
      tags:
      - Rate
    post:
      description: Create Rate for User
      parameters:
      - description: Create Rate Request Body
        in: body
        name: CreateRateRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.CreateRateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create Rate
      tags:
      - Rate
...
  /rate:
    get:
      description: List Account Rate
      parameters:
      - in: query
        name: page_id
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/domain.AccountRateResponse'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List Account Rate
      tags:
      - Rate
  /rooms:
    get:
      description: List Account
      parameters:
      - in: query
        name: page_id
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/domain.ListRoomResponse'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: List Account
      tags:
      - Rooms
    post:
      description: Create Rooms
      parameters:
      - description: create Room Request Body
        in: body
        name: CreateRoomRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.CreateRoomRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.GetRoomResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create Rooms
      tags:
      - Rooms
  /rooms/{room_id}:
    delete:
      description: delete Room
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: delete Room
      tags:
      - Rooms
    get:
      description: Get Room
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.GetRoomResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Room
      tags:
      - Rooms
    put:
      description: update Room
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      - description: update Room Request body
        in: body
        name: UpdateRoomRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateRoomRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.GetRoomResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: update Room
      tags:
      - Rooms
  /rooms/{room_id}/addchat:
    post:
      description: Add Chat Room
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      - description: add chat Room Request body
        in: body
        name: AddChatRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.AddChatRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/domain.GetRoomResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Add Chat Room
      tags:
      - Rooms
  /rooms/{room_id}/members:
    delete:
      description: Remove Account In Rooms
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Remove Account In Rooms
      tags:
      - Rooms
    post:
      description: Add Account In Rooms
      parameters:
      - description: Rooms API wildcard
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Add Account In Rooms
      tags:
      - Rooms
...
swagger: "2.0"